requirements.txt
numpy 
pandas 
mysql-connector-python
-e .
